Fixed bug regarding end loop tokens staying when a loop directly after another one is removed, and got copy token optimization working

+++ src/include/interpreter/Program.hpp
@@ -9,7 +9,7 @@ namespace bs {
 
 	struct Token {
 		char identifier;
-		unsigned int data;
+		long data;
 	};
 
 	struct Program {
